Rename index.js
